requirements.txt
streamlit
beautifulsoup4
requests
pymupdf
python-dotenv
azure-ai-textanalytics
openai==0.28
